Refactor gaussian filter into a new module
Split out all gaussian filter related content into a new module called
gaussian. Also split out associated tests into a new module. Keep an
import in smooth to include the gaussian filter there as well. Update
the import in core to reflect the new location.
