+++ dask_ndfilters/core.py
@@ -9,6 +9,10 @@ from dask_ndfilters.edge import (
     sobel,
 )
 
+from dask_ndfilters.gaussian import (
+    gaussian_filter,
+)
+
 from dask_ndfilters.order import (
     minimum_filter,
     median_filter,
@@ -18,6 +22,5 @@ from dask_ndfilters.order import (
 )
 
 from dask_ndfilters.smooth import (
-    gaussian_filter,
     uniform_filter,
 )
